Update weather api by custom hooks (useWeather)
